Added validation model
Folder for IMECE model contains the old version of the model used to generate the results in Chapter 3.

This commit adds the version of the model used for validation. It is functionally equivalent to the model in the "single pcm" folder, but uses the validation conditions instead of experimental conditions.